Move retryable error wrapping to applyCommand
