+++ pkg/experiment/metastore/metastore_state_poll_compaction_jobs.go
@@ -25,7 +25,6 @@ func (m *Metastore) PollCompactionJobs(_ context.Context, req *compactorv1.PollC
 	_, resp, err := applyCommand[*compactorv1.PollCompactionJobsRequest, *compactorv1.PollCompactionJobsResponse](m.raft, req, m.config.Raft.ApplyTimeout)
 	if err != nil {
 		_ = level.Error(m.logger).Log("msg", "failed to apply poll compaction jobs", "raft_commit_index", m.raft.CommitIndex(), "err", err)
-		return nil, wrapRetryableErrorWithRaftDetails(err, m.raft)
 	}
 	return resp, err
 }
